Added working file for annotations to fix issue #22
I will be working directly in this file to add clarifying annotations to
the vocref-top concepts. This should enable these annotations to be
available in quick iterations without interfering with more substantial
work on -top for now. Since they are just annotations it will be
relatively straightforward to refactor/merge when ready.